+++ DespicableGame/DespicableGame/DespicableGame/GameStates/EtatMenu.cs
@@ -17,9 +17,11 @@ namespace DespicableGame.GameStates
         protected ContentManager content;
         protected InputHandler input;
         private bool exit = false;
-        private readonly int NB_OPTION = 3;
+        private readonly int NB_OPTION = 2;
         private int optionSelectionner = 0;
         private string[] textesMenu;
+        private bool partiePerdu = false;
+        private bool partieGagner = false;
         private SoundEffect menuMusique;
         private SoundEffectInstance menuMusiqueInstance;
 
@@ -29,8 +31,7 @@ namespace DespicableGame.GameStates
             content = _content;
             textesMenu = new string[NB_OPTION];
             textesMenu[0] = "Play";
-            textesMenu[1] = "Classement";
-            textesMenu[2] = "Exit";
+            textesMenu[1] = "Exit";
             input = DespicableGame.input;
 
       
@@ -118,12 +119,6 @@ namespace DespicableGame.GameStates
 
             if (optionSelectionner == 1)
             {
-                DespicableGame.etatDeJeu = new EtatClassement();
-                DespicableGame.etatDeJeu.LoadContent(content);
-            }
-
-            if (optionSelectionner == 2)
-            {
                 exit = true;
             }
         }
@@ -132,10 +127,26 @@ namespace DespicableGame.GameStates
             return exit;
         }
 
+        public void PartiePerdu()
+        {
+            partiePerdu = true;
+        }
+
+        public void PartieGagner()
+        {
+            partieGagner = true;
+        }
+
         public void Draw(SpriteBatch _spriteBatch)
         {
             _spriteBatch.Draw(content.Load<Texture2D>("Sprites\\titleScreen"), Vector2.Zero, null, Color.White, 0f, Vector2.Zero, 1f, SpriteEffects.None, 0f);
             Color couleurTexte;
+
+            if(partiePerdu)
+                _spriteBatch.DrawString(content.Load<SpriteFont>("Font\\MainFont"), "Game Over", new Vector2(550, 290), Color.Red);
+
+            if (partieGagner)
+                _spriteBatch.DrawString(content.Load<SpriteFont>("Font\\MainFont"), "Congratulations! You are now a Pokemon Master!", new Vector2(100, 290), Color.Gold);
 
             for (int i = 0; i < NB_OPTION; i++)
             {
