+++ DespicableGame/DespicableGame/DespicableGame/GameStates/EtatPause.cs
@@ -11,12 +11,11 @@ namespace DespicableGame.GameStates
 {
     class EtatPause : EtatJeu
     {
-
         private EtatPartieEnCours dernierePartieEnCours;
         protected ContentManager content;
         protected InputHandler input;
         private bool exit = false;
-        private readonly int NB_OPTION = 3;
+        private readonly int NB_OPTION = 2;
         private int optionSelectionner = 0;
         private string[] textesMenu;
 
@@ -25,8 +24,7 @@ namespace DespicableGame.GameStates
             content = _content;
             textesMenu = new string[NB_OPTION];
             textesMenu[0] = "Resume";
-            textesMenu[1] = "Retour au menu";
-            textesMenu[2] = "Quitter";
+            textesMenu[1] = "Exit";
             input = DespicableGame.input;
         }
 
@@ -110,13 +108,6 @@ namespace DespicableGame.GameStates
 
             if (optionSelectionner == 1)
             {
-                LevelLoader.Recommencer();
-                DespicableGame.etatDeJeu = new EtatMenu();
-                DespicableGame.etatDeJeu.LoadContent(content);
-            }
-
-            if (optionSelectionner == 2)
-            {
                 exit = true;
             }
         }
@@ -130,7 +121,7 @@ namespace DespicableGame.GameStates
             
             dernierePartieEnCours.Draw(_spriteBatch);
             Color couleurTexte;
-         
+
             for (int i = 0; i < NB_OPTION; i++)
             {
                 couleurTexte = Color.Gray;
