--- DespicableGame/DespicableGame/DespicableGame/Score.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace DespicableGame
-{
-    /// <summary>
-    /// Structure that defines a score with
-    /// the value of the score and the name associated
-    /// </summary>
-    public struct Score
-    {
-        public string name;
-        public int score;
-    }
-}
